[BoxLayout/Grid] translate by integers when scrolling
Translate the actor by whole integers when scrolling, to prevent children
being drawn on non-pixel boundaries.

Fixes bug 4414.

+++ nbtk/nbtk-box-layout.c
@@ -580,7 +580,7 @@ nbtk_box_layout_apply_transform (ClutterActor *a, CoglMatrix *m)
   else
     y = 0;
 
-  cogl_matrix_translate (m , -x, -y, 0);
+  cogl_matrix_translate (m, (int) -x, (int) -y, 0);
 }
